Add logic to munge data and trim sections for fs01 into correct form.
This is to deal with the fact that fs01 has pre-scan turned on, and its detector and data sections are incorrect.
Turning off pre-scan and appropriately adjusting the header keywords requires a great deal of data validation and effort in low-level software.

This logic should be removed when the fs01 spectral camera is retired.