Increase resources for BANZAI workers
With the change to the CR stage, we should stay well below 4gb per worker. So cap it there. Also increase CPU. Finally, we'd like to upgrade the deployment to have 25 worker replicas so that we can keep up with incoming images as the CR stage takes a bit longer to run than most other stages.