Fixes to Frame Factory to make PTR data work
By default PTR doesn't set their EXTNAME in their image HDUs, or it appears they don't. They are set to COMPRESSED_IMAGE, which gets thrown away when the data is unpacked. This causes issues in the in-memory representation of the data, which expects CCDHDUs to have a name (EXTNAME) attribute. Make sure we set this if it's not found, so that processing can complete.