Adds an execution point in main.py
It wasn't clear to me how to execute any banzai functionality without
writing a script. I later realized the entrypoints are listed in
pyproject.toml so meant to be called with an external tool such as
poetry or uv.

Regardless, this PR adds a simple entrypoint to main that adds a
subcommand matching the name of the function to be called. For example,
`python main.py create_db` would be equivalent to `uv run
banzai_create_db`. The changes from `parse_args` to `parse_known_args`
is because argparse by default chokes on extra arguments (which the top
level parser uses) so this allows the individual functions to discard
the extra argument passed to them.

Take it or leave it, this is just a quality of life update.